Support for lex file and output path
Seperated the Gulp options from the Jison options to add support for
reading a lex file and defining the output path in case the input name
doesn't match.

Now the grammar is properly constructed so that Jison is capable of
parsing tokens and generate a working parser and lexer with it.