Added support for generating a parser from JSON.
This commit added support for generating parsers from a grammar provided
in JSON format rather than just .jison format.

Since a .jison grammar will never be valid JSON value, the plugin now
checks if the input is JSON parseable, and if it is, treats it as JSON.
Otherwise it will revert to its previous behaviour.